Remove unused emitter call types and improve INS_tail_i_jmp handling (#59495)
The size estimation for INS_tail_i_jmp instructions was working by
accident: we were always regarding these as requiring rex.w
prefixes, even when they did not. Fix the logic so that rex.w is not
considered required for 8-byte tail jumps, but add it manually for the
forms where it is really necessary.

There were also a lot of unused emitter call types that I have removed.