[mono][interp] Fix unboxing during CALLI (#59747)
* [tests] Re-enable test

* [interp] Fix unboxing during CALLI

On interp we were storing function pointers as a simple InterpMethod. The problem is that, when loading an InterpMethod with a ldvirtftn opcode, the resulting method might need to have its this pointer unboxed, so this information needs to be embedded into the function pointer. We achieve this by tagging the InterpMethod whether we need to perform unboxing as part of calling it.