Share TypedReference managed code between CoreCLR and Mono  (#60636)
* [mono] Rename common TypedReference fields to the same names as CoreCLR

* Share TypedReference managed code between CoreCLR and Mono

The field layout is different. Although both have `_value` as a pointer to the
value (and this PR updates mono to use a `ByReference<byte>` for that pointer,
same as CoreCLR), and a `_type` which is a type handle (`MonoClass` in mono on
the native side), in addition, mono also has a RuntimeTypeHandle (`MonoType` in
mono on the native side) that is used to help the JIT.

But the managed code is identical and only uses `_value` and `_type`.