Refactor Regex FindFirstChar (#60822)
* Refactor Regexcompiler/Emitter.FindFirstChar

These have become unwieldy, with multiple large code blocks for finding the first character at which to perform a full match, and we'll be adding more soon.  This simply splits each of those out into their own method.  No functional changes.

* Use switch in RegexInterpreter.FindFirstChar

We don't need to re-evaluate which mechanism to use on every FindFirstChar call... we can do so once in the ctor and then just switch on that in the actual FindFirstChar call.