Do not explicitly pass type to `VNForMapStore` (#61882)
* Fix type checks in fgValueNumberArrIndexAssign

The code was trying to obtain the value of the location to store
to from the updated map for the element. This is simply incorrect
for all but the trivial case of an empty field sequence.

* Do not explicitly pass type to VNForMapStore

It must always be equal to the type of the map being updated,
not having redundancy eliminates the possibility for mistakes.