Fix Task test which only fails in VS (dotnet/linker#2453)
The problem is that we run the task in a way which never happens in SDK. In SDK we always specify ToolPath/ToolExe values, but the test doesn't. The default implementation of these values rely on DOTNET_HOST_PATH which is not set when running from VS.

The fix calculates path to dotnet.exe from the test itself - it's not the most elegant way, but for now should always work as we only support running the tests on .NET Core (NOT on .NET Framework).

Commit migrated from https://github.com/dotnet/linker/commit/9996319f2a619c2911b02c164b4d6b1d20e29a39