Fix case where we fail to commit the start of a new region, but don't clean up the entry in the region allocator. (#63577)
The region itself though looks like it is on a region free list (heap_segment_allocated is null), so we crash when we try to remove it from its free list, as its containing_free_list is null.

Fix is to delete the region in the region allocator's map, i.e. cause it to be marked as free.