Fix address mode for `IconHandle + Icon` (#63608)
For a NativeAOT change that I'm working on, I'm seeing inefficient address mode getting generated for trees in the shape of `IconHandle + Icon`. The logic I'm touching is swapping this to `Icon + IconHandle` and that makes us miss the if check on the following line because `!op2->IsIntCnsFitsInI32`. That makes us miss generating this tree as `[reg + offset]`. Submitting this as a separate pull request because my actual change will likely need a JitInterface GUID update. This should be no diff for non-AOT scenarios.