Reconfigure the field sequences for statics (#62632)
* Reconfigure the field sequences for statics

To make the "boxed static" and, in the future, "shared" cases identical
to simple ones.

It used to be that we attached the field sequence to the
"inner" constant, and then added the "#FirstElem" pseudo-field for the
outer offset. This was causing complexity and mistakes in the pattern
matching performed by "IsFieldAddr".

The solution is straightforward: always use the outer offset for the sequence.
The much simpler pattern is easier and cheaper to recognize, with less room for
false positives.

* Drop support for mutable RVA statics

They were broken on 32 bit already, and this change does not
break them on 64 bit, but we'll reserve the right to do that
in the future as the scenario appears to be non-existent.