Improve Regex UpdateBumpalong optimization for non-atomic and lazy loops (#63398)
* Improve handling of UpdateBumpalong for non-atomic loops

For atomic greedy loops, UpdateBumpalong is serving its purpose: upon consuming as much as possible for the loop, base.runtextpos is set to that position so that the next time FindFirstChar runs, it'll start from at least that location.  However, for non-atomic greedy loops, with base.runtextpos being set to the ending position after each backtracking, we end up inadvertently voiding any benefits of the UpdateBumpalong, as we end up overwriting the further position with the shorter position.  A simple tweak to that, only setting the position if it's greater, yields significant benefits, in particular when there's no match.

* Add more tests for lazy loops

These just duplicate the greedy loop tests and tweak them for lazy.

* Insert UpdateBumpalong for lazy loops as well

* Address PR feedback